Move icon lib into proper file
