+++ omnilib/prototypes/locale.lua
@@ -3,9 +3,6 @@ local hierarchy = require('locale-hierarchy')
 local lib = { -- We'll return this as omni.locale
     partial = {}
 } 
-local icon = { -- We'll return this as omni.locale
-    partial = {}
-} 
 
 -- All of the below is sourced from the code and methods of TheRustyKnife's locale resolver.
 -- Many thanks to the hours he spent screaming so I don't have to
@@ -386,83 +383,6 @@ function lib.of(prototype, prototype_type)
     return resolver(prototype, locale_type)
 end
 
--- Icons
-function icon.of_generic(prototype, silent)
-    --- Get icons for the given prototype, assuming it's in the generic format.
-    if prototype.icons then
-        local icons = {}
-        for i, icon in pairs(prototype.icons) do
-            if icon.icon and icon.icon_size then icons[i] = icon
-            else
-                local icon_size = icon.icon_size or prototype.icon_size
-                if not icon_size or not icon.icon then
-                    if silent then
-                        return nil
-                    end
-                    error(("%s/%s doesn't specify icons correctly"):format(prototype.type, prototype.name))
-                end                
-                local new = {}
-                for k, v in pairs(icon) do
-                    new[k] = v
-                end
-                new.icon_size = icon_size
-                icons[i] = new
-            end
-        end
-        return icons
-    end
-    if not prototype.icon or not prototype.icon_size then
-        if silent then
-            return nil
-        end
-        error(("%s/%s doesn't specify icons correctly"):format(prototype.type, prototype.name))
-    end
-    return {{
-        icon = prototype.icon,
-        icon_size = prototype.icon_size
-    }}
-end
-
-function icon.of_recipe(prototype, silent)
-    --- Get icons for the given recipe prototype.
-    local icons = icon.of_generic(prototype, true)
-    if icons then return icons; end
-
-    local product
-    if prototype.normal ~= nil then product = lib.partial.get_main_product(prototype.normal)
-    else product = lib.partial.get_main_product(prototype); end
-
-    if not product then
-        if silent then
-            return nil
-        end
-        error(("%s/%s doesn't specify icons correctly"):format(prototype.type, prototype.name))
-    end
-
-    return product and icon.of(product.name, product.type, silent)
-end
-
-
-function icon.of(prototype, ptype, silent)
-    --- Get the icons of the given prototype.
-    if type(ptype) == 'string' then
-        prototype = lib.find(prototype, ptype, silent)
-    elseif prototype == nil then
-        if silent then
-            return nil 
-        end
-        error("Can't get icons of nil prototype")
-    elseif type(ptype) == 'boolean' then -- wrong arg order
-        silent = ptype 
-    end
-
-    if lib.inherits(prototype.type, 'recipe') then
-        return icon.of_recipe(prototype, silent)
-    else
-        return icon.of_generic(prototype, silent)
-    end
-end
-
 lib.custom_name = function(prototype, name_key, ...)
     if not name_key:find("%.") then
         name_key = lib.inherits(prototype.type, lib.localised_types).."-name."..name_key
@@ -485,4 +405,3 @@ lib.custom_description = function(prototype, desc_key, ...)
     }
 end
 omni.locale = lib
-omni.icon = icon
